configurator: Extract cv to separate crate
