+++ configurator/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+rustmas-cv = { path = "../cv" }
 rustmas-light-client = { path = "../light-client", default-features = false }
 lightfx = { path = "../lightfx" }
 
@@ -11,7 +12,6 @@ csv = "1.1.6"
 clap = { version = "4.0.8", features = ["derive"] }
 indicatif = "0.17.1"
 itertools = "0.10.5"
-opencv = "0.70"
 serde = { version = "1.0", features = ["derive"] }
 tokio = { version = "1", features = ["full"] }
